requirements.txt
PyYAML==5.3.1
apache-airflow[crypto,celery,postgres,hive,jdbc,ssh]==1.10.10
boto3==1.14.18
google-cloud-bigquery==1.25.0
google-oauth==1.0.1
dateparser==0.7.6
cloudpickle==1.4.1
dask[complete]==2.20.0
distributed==2.20.0
google-api-python-client==1.9.3
bigquery-schema-generator==1.0
tweepy==3.8.0
GetOldTweets3==0.0.11
